% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_EquiEquate.R
\name{bh_linear_equate}
\alias{bh_linear_equate}
\title{Perform Braun-Holland Linear Equating}
\usage{
bh_linear_equate(minx, maxx, incx, fxs, miny, maxy, incy, gys)
}
\description{
Calculates the slope and intercept for Braun-Holland linear equating based on
synthetic densities. This is an R translation of the `BH_LinEq` function.
}
