% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{biv_exp_polynomial}
\alias{biv_exp_polynomial}
\title{Bivariate Exponential Polynomial}
\usage{
biv_exp_polynomial(bivar_list, x, y)
}
\description{
Translates `BivExpPolynomial`. This version is NOT vectorized.
}
