% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{biv_exp_polynomial_vec}
\alias{biv_exp_polynomial_vec}
\title{Vectorized Bivariate Exponential Polynomial Calculation}
\usage{
biv_exp_polynomial_vec(bivar, x, y)
}
\description{
The core computational engine for the log-linear model's density function.
This version is highly optimized for speed using matrix operations.
}
