% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{bootstrap_univariate}
\alias{bootstrap_univariate}
\title{Generate a Non-Parametric Bootstrap Sample for a Univariate Distribution}
\usage{
bootstrap_univariate(n, freq)
}
\arguments{
\item{n}{The total sample size.}

\item{freq}{A numeric vector of the original frequency distribution.}
}
\value{
A numeric vector representing the new bootstrap frequency distribution.
}
\description{
Creates a bootstrap sample from an observed univariate frequency distribution.
This is an R replacement for the C function `Boot_USTATS`.
}
\details{
This function uses sampling from a multinomial distribution, which is an
efficient way to perform non-parametric bootstrapping from frequency data.
}
