% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaBinomial.R
\name{calc_kurt}
\alias{calc_kurt}
\title{Calculate Kurtosis from Beta Parameters}
\usage{
calc_kurt(alpha, beta)
}
\arguments{
\item{alpha}{The first shape parameter (alpha > 0).}

\item{beta}{The second shape parameter (beta > 0).}
}
\value{
The numeric value of the kurtosis.
}
\description{
Calculates the kurtosis of a beta distribution given its two shape parameters.
This is an R translation of `CalcKurt`.
}
\author{
B. A. Hanson (Original C code)
}
