% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_helpers.R
\name{calculate_icc}
\alias{calculate_icc}
\title{Calculate Item Characteristic Curves (ICCs)}
\usage{
calculate_icc(irt_pars, theta)
}
\arguments{
\item{irt_pars}{A data frame of item parameters (`slope`, `intercept`, `guess`).}

\item{theta}{A numeric vector of ability values.}
}
\value{
A matrix of probabilities, with items in rows and theta points in columns.
}
\description{
A helper function to compute the ICCs from a set of dichotomous item
parameters for a given set of ability values.
}
\keyword{internal}
