% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_EquiEquate.R
\name{chained_equate}
\alias{chained_equate}
\title{Perform Chained Equipercentile Equating for CINEG Design}
\usage{
chained_equate(
  nsx,
  prx1,
  minv,
  maxv,
  incv,
  nsv,
  Gv1,
  miny,
  incy,
  nsy,
  Gy2,
  Gv2
)
}
\description{
Implements chained equipercentile equating as described in Kolen & Brennan (2004).
This is an R translation of `Chained_EE`.
}
