% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{cll_equate_cb}
\alias{cll_equate_cb}
\title{Perform CLL Equating for the Counter-Balanced Design}
\usage{
cll_equate_cb(bivar1, bivar2, wtsx, wtsy)
}
\description{
Perform CLL Equating for the Counter-Balanced Design
}
