% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{cll_equate_neat_chn}
\alias{cll_equate_neat_chn}
\title{Perform CLL Equating for NEAT Chained Design}
\usage{
cll_equate_neat_chn(bivar1, bivar2)
}
\description{
Perform CLL Equating for NEAT Chained Design
}
