% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{cll_equate_neat_chn_rcpp}
\alias{cll_equate_neat_chn_rcpp}
\title{Perform CLL Equating (NEAT Chained) using Rcpp for Speed}
\usage{
cll_equate_neat_chn_rcpp(bivar1, bivar2)
}
\arguments{
\item{bivar1, }{bivar2 The `bivar` objects for the two forms.}
}
\value{
A numeric vector of equated scores.
}
\description{
A high-performance version of `cll_equate_neat_chn` that calls the compiled
C++ functions for the most intensive calculations.
}
