% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{cll_equate_neat_ps}
\alias{cll_equate_neat_ps}
\title{Perform CLL Equating for NEAT Post-Stratification Design}
\usage{
cll_equate_neat_ps(bivar1, bivar2, wts)
}
\description{
Perform CLL Equating for NEAT Post-Stratification Design
}
