% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{cll_equate_sg_cpp}
\alias{cll_equate_sg_cpp}
\title{CLL Single-Group Equating (C++ Accelerated)}
\usage{
cll_equate_sg_cpp(bivar)
}
\arguments{
\item{bivar}{A bivariate model object created by `smooth_bll`.}
}
\value{
A vector of equated Y scores corresponding to the input X scores.
}
\description{
This is the main wrapper function for single-group CLL equating.
}
