% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{cll_marginal_cdf_x_cpp}
\alias{cll_marginal_cdf_x_cpp}
\title{Calculate Marginal CDF for X (C++ Accelerated)}
\usage{
cll_marginal_cdf_x_cpp(x_val, bivar, nc)
}
\description{
Calculates the cumulative distribution function P(X <= x_val).
}
