% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{cll_marginal_cdf_y_cpp}
\alias{cll_marginal_cdf_y_cpp}
\title{Calculate Marginal CDF for Y (C++ Accelerated)}
\usage{
cll_marginal_cdf_y_cpp(y_val, bivar, nc)
}
\description{
Calculates the cumulative distribution function P(Y <= y_val).
}
