% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{cll_marginal_inverse_cdf_x}
\alias{cll_marginal_inverse_cdf_x}
\title{Calculate Inverse Marginal CDF for X}
\usage{
cll_marginal_inverse_cdf_x(p, bivar, nc)
}
\description{
Calculate Inverse Marginal CDF for X
}
