% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{cll_marginal_inverse_cdf_y}
\alias{cll_marginal_inverse_cdf_y}
\title{Calculate Inverse Marginal CDF for Y}
\usage{
cll_marginal_inverse_cdf_y(p, bivar, nc)
}
\description{
Calculate Inverse Marginal CDF for Y
}
