% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{cll_marginal_inverse_cdf_y_cpp}
\alias{cll_marginal_inverse_cdf_y_cpp}
\title{Calculate Inverse Marginal CDF for Y (C++ Accelerated)}
\usage{
cll_marginal_inverse_cdf_y_cpp(p, bivar, nc)
}
\description{
Finds the score y such that P(Y <= y) = p.
}
