% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{cll_marginal_pdf_x}
\alias{cll_marginal_pdf_x}
\title{Calculate Marginal PDF for X from a Bivariate Distribution}
\usage{
cll_marginal_pdf_x(x_vec, bivar, nc)
}
\description{
Calculate Marginal PDF for X from a Bivariate Distribution
}
