% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{cll_marginal_pdf_y}
\alias{cll_marginal_pdf_y}
\title{Calculate Marginal PDF for Y from a Bivariate Distribution}
\usage{
cll_marginal_pdf_y(y_vec, bivar, nc)
}
\description{
Calculate Marginal PDF for Y from a Bivariate Distribution
}
