% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_Equate.R
\name{compute_c_matrix}
\alias{compute_c_matrix}
\title{Calculate the C Matrix for Delta Method Standard Errors}
\usage{
compute_c_matrix(n_persons, design_matrix, probabilities)
}
\arguments{
\item{n_persons}{Total number of examinees.}

\item{design_matrix}{The design matrix (B) from the log-linear model.}

\item{probabilities}{The vector of smoothed probabilities (p) from the log-linear model.}
}
\value{
The C matrix.
}
\description{
Computes the C matrix such that C * t(C) is the variance-covariance
matrix of the log-linear probabilities. R translation of `ComputeCmatrixGen`.
}
