% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_EquiEquate.R
\name{cond_biv_dist}
\alias{cond_biv_dist}
\title{Calculate a Conditional Bivariate Distribution}
\usage{
cond_biv_dist(bxv, hv)
}
\arguments{
\item{bxv}{A matrix of bivariate relative frequencies (x by v).}

\item{hv}{A vector of the marginal relative frequencies for v.}
}
\value{
A matrix representing the conditional distribution P(X|V).
}
\description{
Calculates the conditional distribution of one variable given another, i.e., P(X|V).
This is an R translation of the `CondBivDist` function.
}
