% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubicSpline.R
\name{cubic_poly}
\alias{cubic_poly}
\title{Evaluate a Cubic Polynomial}
\usage{
cubic_poly(x_left, a, b, c, d, xvalue)
}
\arguments{
\item{x_left}{The left boundary of the interval for the polynomial.}

\item{a}{The constant coefficient.}

\item{b}{The linear coefficient.}

\item{c}{The quadratic coefficient.}

\item{d}{The cubic coefficient.}

\item{xvalue}{The x-value at which to evaluate the polynomial.}
}
\value{
The numeric value of the polynomial at `xvalue`.
}
\description{
Evaluates a cubic polynomial of the form:
`f(x) = a + b*(x-x_left) + c*(x-x_left)^2 + d*(x-x_left)^3`
for a given `xvalue`.
}
