% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERutilities.R
\name{cum_rel_freqs}
\alias{cum_rel_freqs}
\title{Compute Cumulative Relative Frequencies}
\usage{
cum_rel_freqs(rel_freq)
}
\arguments{
\item{rel_freq}{A numeric vector of relative frequencies.}
}
\value{
A numeric vector of cumulative relative frequencies.
}
\description{
Computes the cumulative relative frequencies from a vector of relative
frequencies. This is an R replacement for the C function `cum_rel_freqs`.
}
