% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLinear.R
\name{design_matrix}
\alias{design_matrix}
\title{Create a Design Matrix for Log-Linear Models}
\usage{
design_matrix(
  nsu,
  minu,
  incu,
  nsv = 0,
  minv = 0,
  incv = 0,
  cu,
  cv = 0,
  cuv = 0,
  cpm = NULL,
  scale = FALSE
)
}
\value{
A list containing the raw design matrix (`B_raw`), the scaled design
  matrix (`B`), and the scaling attributes (`B_center`, `B_scale`).
}
\description{
Creates a design matrix and returns the scaling attributes (center and scale)
used during its creation, which are necessary for unscaling the results.
}
