% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERutilities.R
\name{equated_ss}
\alias{equated_ss}
\title{Get Equated Scale Scores}
\usage{
equated_ss(eraw, yct, lprss = NULL, hprss = NULL, round_digits = 0)
}
\arguments{
\item{eraw}{A numeric vector of equated raw scores.}

\item{yct}{A data frame or matrix with two columns: the first for the raw
scores on the base form (Y), and the second for the corresponding scale scores.}

\item{lprss}{The lowest possible rounded scale score (for clipping).}

\item{hprss}{The highest possible rounded scale score (for clipping).}

\item{round_digits}{An integer specifying the number of decimal places to
round the final scale scores to. If NULL, no rounding is performed.}
}
\value{
A list containing the unrounded (`essu`) and rounded (`essr`)
  equated scale scores.
}
\description{
Converts equated raw scores to a scale score metric using a provided
raw-to-scale score conversion table. This function uses linear interpolation
for equated raw scores that fall between the points in the conversion table.
This is an R translation of the `Equated_ss` function.
}
