% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERutilities.R
\name{er_ludcmp}
\alias{er_ludcmp}
\title{Perform LU Decomposition of a Matrix}
\usage{
er_ludcmp(a)
}
\arguments{
\item{a}{A square numeric matrix.}
}
\value{
A list containing the LU-decomposed matrix and pivot information,
  compatible with `er_lubksb`. In R, this is often handled by `solve` or
  functions from the `Matrix` package. For simplicity, we can use `solve`
  as a high-level equivalent.
}
\description{
This function serves as a wrapper for R's LU decomposition capabilities,
primarily for compatibility with the C code's structure. It computes the
LU decomposition of a square matrix.
}
\details{
The original C function `er_ludcmp` performs LU decomposition with partial
pivoting and modifies the input matrix `a` to store the L and U factors.
This R version will simply return the result of an LU decomposition, which
can then be used by `er_lubksb`. For most R applications, one would use
`solve()` directly.
}
\seealso{
\code{\link{er_lubksb}}, \code{\link{solve}}
}
