% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERutilities.R
\name{er_rtsafe}
\alias{er_rtsafe}
\title{Find the Root of a One-Dimensional Function}
\usage{
er_rtsafe(f, interval)
}
\arguments{
\item{f}{The function for which the root is sought.}

\item{interval}{A vector containing the two endpoints of the interval to search.}
}
\value{
The x-value of the root.
}
\description{
Finds the root of a function within a given interval. This is a wrapper for
R's `uniroot` function, replacing the C library's `er_rtsafe`.
}
