% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{exp_polynomial}
\alias{exp_polynomial}
\title{Evaluate the Exponential of a Polynomial (Vectorized)}
\usage{
exp_polynomial(x_vec, params)
}
\description{
The core component of the log-linear model's density function.
Translates `ExpPolynomial`. This version is vectorized to work with `integrate`.
}
