% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_EquiEquate.R
\name{fe_mfe_equate}
\alias{fe_mfe_equate}
\title{Perform Frequency Estimation (FE) or Modified FE (MFE) Equipercentile Equating}
\usage{
fe_mfe_equate(
  w1,
  internal,
  bxvin,
  byvin,
  rv1,
  rv2,
  minx,
  maxx,
  incx,
  miny,
  maxy,
  incy
)
}
\description{
The main function for FE or MFE equipercentile equating in a CINEG design.
This is an R translation of `FEorMFE_EE`.
}
