% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_Equate.R
\name{find_optimal_h}
\alias{find_optimal_h}
\title{Find Optimal Bandwidth (h) for Kernel Equating}
\usage{
find_optimal_h(scores, rel_freq, search_interval = c(0.1, 2))
}
\arguments{
\item{scores}{A numeric vector of the discrete score points.}

\item{rel_freq}{A numeric vector of the relative frequencies for each score.}

\item{search_interval}{A vector of length 2 defining the search interval for `h`.}
}
\value{
The optimal bandwidth value `h`.
}
\description{
Finds the optimal bandwidth `h` by minimizing the penalty function.
This is an R translation of `Optimalh`.
}
