% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{gaussian_quadrature}
\alias{gaussian_quadrature}
\title{Numerical Integration using R's `integrate`}
\usage{
gaussian_quadrature(f, lower, upper, ...)
}
\description{
A general-purpose numerical integration function to replace the specific
`GaussianQuadrature` functions in the C code.
}
