% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{gaussian_quadrature_cpp}
\alias{gaussian_quadrature_cpp}
\title{Numerically Integrate Using C++ Backend (Vectorized)}
\usage{
gaussian_quadrature_cpp(bivar, lower_x, upper_x, lower_y, upper_y)
}
\arguments{
\item{bivar}{The bivariate model object.}

\item{lower_x}{Lower integration limit for X.}

\item{upper_x}{Upper integration limit for X.}

\item{lower_y}{Lower integration limit for Y.}

\item{upper_y}{Upper integration limit for Y.}
}
\value{
The value of the double integral.
}
\description{
This function performs a double integral using Gauss-Legendre quadrature.
It calls the C++ function `integrand_fully_vectorized_cpp` to evaluate
the inner function efficiently over a grid.
}
