% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLL_Equate.R
\name{gaussian_quadrature_vec}
\alias{gaussian_quadrature_vec}
\title{Gaussian Quadrature for Numerical Integration (Fixed-Point)}
\usage{
gaussian_quadrature_vec(f, lower, upper, ...)
}
\description{
Approximates a definite integral using a 20-point Legendre-Gauss fixed-point rule.
}
