% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_recipe.R
\name{get_method_options}
\alias{get_method_options}
\title{Get Master List of Default Method Options}
\usage{
get_method_options()
}
\value{
A list containing the complete set of default options.
}
\description{
Provides a master list of default arguments for all equating methods and their
options. This function serves as a single source of truth for default values
used throughout the package, primarily by `add_method()`.
}
\keyword{internal}
