% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utilities.R
\name{get_range}
\alias{get_range}
\title{# Get the range automatically and print the summary
range_info <- get_range(sample_ctabs, verbose = TRUE)}
\usage{
get_range(
  ctabs,
  form_name = NULL,
  min_score = NULL,
  max_score = NULL,
  inc = NULL,
  verbose = TRUE
)
}
\description{
# Get the range with a manual override for the minimum score
range_override <- get_range(sample_ctabs, min_score = 0)
}
