% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_helpers.R
\name{get_theta_dist}
\alias{get_theta_dist}
\title{Helper to get theta distribution vector}
\usage{
get_theta_dist(theta_dist_option, theta)
}
\description{
Helper to get theta distribution vector
}
\keyword{internal}
