% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_EquiEquate.R
\name{interpolate}
\alias{interpolate}
\title{Linearly Interpolate or Extrapolate a Value}
\usage{
interpolate(x, ns, f)
}
\arguments{
\item{x}{The numeric point at which to interpolate the value.}

\item{ns}{The number of score categories (length of `f`).}

\item{f}{A numeric vector of function values at points `0, 1, ..., ns-1`.}
}
\value{
The interpolated or extrapolated value.
}
\description{
Interpolates or extrapolates the value of a function `f` at point `x`.
The function `f` is defined by a vector of values at integer points
from 0 to `ns-1`. This is an R translation of the `interpolate` function.
}
