% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubicSpline.R
\name{inverse_post_smooth}
\alias{inverse_post_smooth}
\title{Compute the Inverse of the Post-Smoothed Equating Function}
\usage{
inverse_post_smooth(yvalues, xvalues, dxi, s, ylow, yhigh, kx, vectX)
}
\arguments{
\item{yvalues}{Original y-scores (knots).}

\item{xvalues}{Original equated x-scores.}

\item{dxi}{Standard errors of equated scores.}

\item{s}{Smoothing parameter.}

\item{ylow, yhigh}{Indices for the spline fitting range.}

\item{kx}{Maximum score on the X scale.}

\item{vectX}{The x-values at which to evaluate the inverse.}
}
\value{
A vector of y-values from the inverse function.
}
\description{
Computes the inverse of the complete post-smoothed equating function,
including the linear tail sections.
}
