% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubicSpline.R
\name{inverse_sspline}
\alias{inverse_sspline}
\title{Compute the Inverse of a Full Cubic Spline Function}
\usage{
inverse_sspline(ynodes, cmat, vectX)
}
\arguments{
\item{ynodes}{The original y-nodes (knots) of the spline.}

\item{cmat}{The coefficient matrix of the spline.}

\item{vectX}{The x-values at which to evaluate the inverse function.}
}
\value{
A vector of y-values from the inverse function.
}
\description{
Computes the inverse of a complete cubic spline function at specified points.
}
