% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_helpers.R
\name{irt}
\alias{irt}
\title{IRT Equating Wrapper}
\usage{
irt(forms, design, type, eq, title, ...)
}
\arguments{
\item{forms}{A character vector of length two indicating the forms to be equated.}

\item{design}{A character string for the equating design (e.g., "cg").}

\item{type}{A character string specifying the IRT equating type
(e.g., "true_score").}

\item{eq}{The `equate_recipe` object containing all forms and methods.}

\item{title}{The unique title of the method being run, used to retrieve
the correct options from the `eq` object.}

\item{...}{Additional arguments, for forward compatibility.}
}
\description{
This function serves as a wrapper for different types of IRT equating,
such as true score and observed score equating. It retrieves the necessary
parameters from the `equate_recipe` object and calls the appropriate
low-level IRT equating engine.
}
\keyword{internal}
