% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_helpers.R
\name{irt_observed_score_equate}
\alias{irt_observed_score_equate}
\title{Perform IRT Observed Score Equating}
\usage{
irt_observed_score_equate(
  irt_pars_x,
  irt_pars_y,
  theta,
  eq,
  forms,
  design,
  method_options,
  min_score = NULL,
  max_score = NULL
)
}
\arguments{
\item{irt_pars_x}{A data frame of IRT item parameters for Form X (on the same scale as Y).}

\item{irt_pars_y}{A data frame of IRT item parameters for Form Y.}

\item{theta}{A numeric vector for the `theta` grid.}

\item{eq}{The `equate_recipe` object.}

\item{forms}{A character vector of the two forms being equated.}

\item{design}{The equating design.}

\item{method_options}{A list of options for the method.}

\item{min_score}{Optional. The minimum score for the score scale. If NULL,
it's derived from the `eq` object.}

\item{max_score}{Optional. The maximum score for the score scale.}
}
\description{
Performs IRT observed score equating for SG and CNEG designs.
}
\details{
This procedure derives the observed score distributions for each form from
the IRT model and a population ability distribution. For the CNEG design, it
creates a synthetic population based on the two groups. It then performs an
equipercentile equating on these two model-based distributions.
}
\keyword{internal}
