% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLinear.R
\name{iteration}
\alias{iteration}
\title{Main Iterative Algorithm for Log-Linear Fitting}
\usage{
iteration(B, B_raw, nct, N, max_nit = 100, crit = 1e-05)
}
\description{
The Newton-Raphson algorithm to fit the log-linear model.
}
