% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_Equate.R
\name{kernel_equate}
\alias{kernel_equate}
\title{Perform Kernel Equipercentile Equating}
\usage{
kernel_equate(scores_x, rel_freq_x, hx, scores_y, rel_freq_y, hy)
}
\arguments{
\item{scores_x, }{scores_y Numeric vectors of score points for Form X and Y.}

\item{rel_freq_x, }{rel_freq_y Numeric vectors of relative frequencies for Form X and Y.}

\item{hx, }{hy Optimal bandwidth parameters for Form X and Y.}
}
\value{
A numeric vector of equated Y scores for each score on Form X.
}
\description{
This is the core kernel equating function. It takes two distributions
and finds the equated scores. R translation of `KernelEquate`.
}
