% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_Equate.R
\name{kernel_inverse_cdf}
\alias{kernel_inverse_cdf}
\title{Inverse of the Kernel Smoothed CDF}
\usage{
kernel_inverse_cdf(p, scores, rel_freq, hx, tol = 1e-08)
}
\arguments{
\item{p}{A numeric value (probability) between 0 and 1.}

\item{scores}{A numeric vector of the discrete score points.}

\item{rel_freq}{A numeric vector of the relative frequencies for each score.}

\item{hx}{A numeric value for the bandwidth parameter.}
}
\value{
The score `x` corresponding to the cumulative probability `p`.
}
\description{
Finds the score `x` that corresponds to a given cumulative probability `p`.
}
