% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_Equate.R
\name{kernel_pdf_derivative}
\alias{kernel_pdf_derivative}
\title{First Derivative of the Kernel PDF}
\usage{
kernel_pdf_derivative(x, scores, rel_freq, hx)
}
\description{
R translation of `KernelPdfDerivative`.
}
