% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_NoSmooth.R
\name{linear_equate_observed_ci}
\alias{linear_equate_observed_ci}
\title{Calculate Synthetic Population Moments for Linear Equating (CINEG)}
\usage{
linear_equate_observed_ci(
  mnx1,
  sdx1,
  mnv1,
  sdv1,
  mny2,
  sdy2,
  mnv2,
  sdv2,
  w1,
  mean_only,
  gamma1,
  gamma2
)
}
\description{
A helper function for CINEG linear equating. It calculates the means and
standard deviations of the synthetic population, as well as the slope and
intercept of the linear equating function. This is an R translation of the
`CI_LinObsEq` function.
}
