% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetaBinomial.R
\name{lr_chi_sqr}
\alias{lr_chi_sqr}
\title{Likelihood Ratio Chi-Square}
\usage{
lr_chi_sqr(raw_counts, fit_counts)
}
\arguments{
\item{raw_counts}{A vector of observed frequencies.}

\item{fit_counts}{A vector of fitted frequencies from the model.}
}
\value{
The numeric chi-square value.
}
\description{
Computes the likelihood ratio chi-squared statistic to assess goodness-of-fit.
This is an R translation of `LRChiSqr`.
}
