% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_EquiEquate.R
\name{mix_smooth}
\alias{mix_smooth}
\title{Smooth a Bivariate Distribution by Mixing with a Uniform Distribution}
\usage{
mix_smooth(bxv, unimix = 1e-10)
}
\arguments{
\item{bxv}{A matrix of bivariate relative frequencies (x by v).}

\item{unimix}{The mixing proportion for the uniform distribution.}
}
\value{
A list containing the smoothed bivariate matrix (`bxv_smoothed`),
  and the smoothed marginal distributions for x (`fx`) and v (`hv`).
}
\description{
Smooths a bivariate distribution to ensure no cell has a zero probability.
This is an R translation of the `MixSmooth` function.
}
