% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_EquiEquate.R
\name{mod_cond_biv_dist}
\alias{mod_cond_biv_dist}
\title{Modify a Conditional Bivariate Distribution for MFE Equating}
\usage{
mod_cond_biv_dist(internal, nsv, nsx, rv1, rv2, muv1, muv2, bxv)
}
\arguments{
\item{internal}{A logical, TRUE if the anchor is internal, FALSE if external.}

\item{nsv, nsx}{Number of score categories for v and x.}

\item{rv1, rv2}{Reliability of the anchor test in populations 1 and 2.}

\item{muv1, muv2}{Mean of the anchor test in populations 1 and 2.}

\item{bxv}{The conditional distribution P(X|V) to be modified.}
}
\value{
The modified conditional distribution matrix.
}
\description{
For Modified Frequency Estimation (MFE), this function modifies the
conditional distribution of x given v. This is an R translation of
the `ModCondBivDist` function.
}
