% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_Equate.R
\name{partial_f_partial_r}
\alias{partial_f_partial_r}
\title{Calculate Partial Derivatives for Delta Method}
\usage{
partial_f_partial_r(scores, rel_freq, hx)
}
\value{
A matrix of partial derivatives.
}
\description{
Calculates the partial derivative of the kernel CDF F with respect
to the relative frequencies r. R translation of `PartialFPartialr`.
}
