% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utilities.R
\name{print.equate_results}
\alias{print.equate_results}
\title{Print a Summary of Equating Results Object}
\usage{
\method{print}{equate_results}(x, ...)
}
\arguments{
\item{x}{An object of class `equate_results`.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object `x`.
}
\description{
Print a Summary of Equating Results Object
}
