% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_print.R
\name{print_result_summary}
\alias{print_result_summary}
\title{Print a concise summary of a single equating result}
\usage{
print_result_summary(result, recipe, from_form)
}
\description{
Print a concise summary of a single equating result
}
\keyword{internal}
