% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_helpers.R
\name{scale_curve}
\alias{scale_curve}
\title{Perform IRT Scale Transformation using Characteristic Curve Methods}
\usage{
scale_curve(common_items_x, common_items_y, theta, method = "stocking_lord")
}
\arguments{
\item{common_items_x}{Standardized IRT parameters for common items on Form X.}

\item{common_items_y}{Standardized IRT parameters for common items on Form Y.}

\item{theta}{A numeric vector for the `theta` grid.}

\item{method}{Either `"stocking_lord"` or `"haebara"`.}
}
\value{
A list containing the estimated slope `A` and intercept `B`.
}
\description{
Estimates scale transformation constants A (slope) and B (intercept) using
optimization to minimize the difference between item or test characteristic
curves of the common items.
}
\details{
This function uses `stats::optim()` to find the `A` and `B` constants that
minimize a criterion function. The method determines the nature of this
function:
\itemize{
  \item **stocking_lord:** Minimizes the sum of squared differences between the
    **Test** Characteristic Curves (TCCs) of the common items.
  \item **haebara:** Minimizes the sum of squared differences between all
    **Item** Characteristic Curves (ICCs) of the common items.
}
}
\keyword{internal}
