% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utilities.R
\name{summarize_equating_results}
\alias{summarize_equating_results}
\title{Summarize Equating Results}
\usage{
summarize_equating_results(results_object, ...)
}
\arguments{
\item{results_object}{An object of class `equate_results`, the output from `run_equating`.}

\item{...}{Additional arguments (not currently used).}
}
\value{
A nested list where each element corresponds to an equating plan.
  Each plan's element is a list containing the summary components
  (tables and plots) for that specific plan.
}
\description{
Provides a comprehensive summary for each equating plan within an analysis
performed by `run_equating`. It generates comparative tables and plots for
the methods applied within each plan.
}
