% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_EquiEquate.R
\name{synthetic_densities}
\alias{synthetic_densities}
\title{Calculate Synthetic Population Densities}
\usage{
synthetic_densities(w1, internal, bxvin, byvin, rv1, rv2)
}
\arguments{
\item{w1}{Weight for population 1.}

\item{internal}{Logical, TRUE for internal anchor.}

\item{bxvin, byvin}{Bivariate relative frequency matrices for (X,V) and (Y,V).}

\item{rv1, rv2}{Reliabilities of anchor test V in each population.}
}
\value{
A list containing the synthetic densities `fxs` and `gys`.
}
\description{
Calculates the synthetic population densities for forms X and Y for the CINEG
design. This is an R translation of the `SyntheticDensities` function.
}
