% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_helpers.R
\name{transform_irt_pars}
\alias{transform_irt_pars}
\title{Apply Scale Transformation to IRT Parameters}
\usage{
transform_irt_pars(irt_pars, A, B)
}
\arguments{
\item{irt_pars}{A data frame of standardized IRT parameters.}

\item{A}{The slope of the transformation.}

\item{B}{The intercept of the transformation.}
}
\value{
A data frame with the transformed parameters.
}
\description{
Transforms a set of IRT parameters from one scale to another using the
provided A and B constants, based on the slope-intercept parameterization.
}
\keyword{internal}
