% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_print.R
\name{translate_title}
\alias{translate_title}
\title{Translate a method title into a human-readable string}
\usage{
translate_title(title)
}
\description{
Translate a method title into a human-readable string
}
\keyword{internal}
