% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLinear.R
\name{unscale_beta}
\alias{unscale_beta}
\title{Convert Scaled Beta Coefficients to Unscaled Metric}
\usage{
unscale_beta(Beta_scaled, ap_scaled, B_center, B_scale)
}
\value{
A list containing the unscaled Beta coefficients and intercept.
}
\description{
Transforms the log-linear model's Beta coefficients and intercept (ap)
from a scaled metric back to the original, unscaled metric.
}
