% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogLinear.R
\name{design_matrix}
\alias{design_matrix}
\title{Create a Design Matrix for Log-Linear Models}
\usage{
design_matrix(
  nsu,
  minu,
  incu,
  nsv = 0,
  minv = 0,
  incv = 0,
  cu,
  cv = 0,
  cuv = 0,
  cpm = NULL,
  scale = FALSE
)
}
\arguments{
\item{nsu, }{nsv Number of score categories for variables u and v.}

\item{minu, }{minv, incu, incv Score range and increment for u and v.}

\item{cu, }{cv, cuv Degrees for polynomials and cross-products.}

\item{cpm}{A matrix specifying the cross-product terms.}

\item{scale}{Logical, if TRUE, scale the design matrix columns.}
}
\value{
A list containing the raw design matrix (`B_raw`) and the
  potentially scaled design matrix (`B`).
}
\description{
Creates a design matrix based on powers of scores for univariate or
bivariate log-linear models.
}
