% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_utilities.R
\name{equate}
\alias{equate}
\title{Internal Equating Dispatcher}
\usage{
equate(forms, method, design, type, eq, title)
}
\arguments{
\item{forms}{A character vector of length two indicating the forms to be equated.}

\item{method}{A single-character code (`"L"`, `"E"`, `"IRT"`) specifying the
primary equating method.}

\item{design}{A character string for the equating design (e.g., "cg").}

\item{type}{A character string specifying the sub-method or calculation type.}

\item{eq}{The `equate_recipe` object containing all forms and methods.}

\item{title}{The unique title of the method being run, used to retrieve
the correct options from the `eq` object.}

\item{boot_type}{The type of bootstrap confidence intervals.}

\item{boot_replications}{The number of bootstrap replications.}
}
\description{
This function is the internal switchboard that calls the appropriate method-specific
wrapper (`linear`, `equipercentile`, `irt`) based on the method specified
in the `equate_recipe`. It is called by `run_equating()` for each method
defined in the recipe's plan.
}
\keyword{internal}
