% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_helpers.R
\name{irt}
\alias{irt}
\title{IRT Equating Wrapper}
\usage{
irt(forms, design, type, eq, title, ...)
}
\arguments{
\item{forms}{A character vector of the two forms being equated.}

\item{design}{A character string for the equating design ("sg" or "cneg").}

\item{type}{A character string specifying the IRT equating type.}

\item{eq}{The `equate_recipe` object.}

\item{title}{The unique title of the method being run.}
}
\description{
This function is the main internal dispatcher for IRT equating. It determines
the equating design, performs scale transformation if necessary (for CNEG),
and then calls the appropriate equating engine (`true_score` or `observed_score`).
}
\keyword{internal}
