% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_helpers.R
\name{irt_coefs}
\alias{irt_coefs}
\title{Standardize IRT Item Parameters}
\usage{
irt_coefs(irt_pars)
}
\arguments{
\item{irt_pars}{The item parameters, which can be a data.frame, a `mirt`
model object, a `TAM` model object, or a list of these.}
}
\value{
A data frame or a list of data frames with standardized item parameters.
}
\description{
A helper function to extract and standardize IRT item parameters from various
sources into a consistent data frame format (`slope`, `intercept`, `guess`).
}
\keyword{internal}
