% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubicSpline.R
\name{post_smooth}
\alias{post_smooth}
\title{Apply Post-Smoothing Using Cubic Splines}
\usage{
post_smooth(xvalues, yvalues, dyi, s, xlow, xhigh, ky, vectX)
}
\arguments{
\item{xvalues}{A numeric vector of raw scores.}

\item{yvalues}{A numeric vector of equated scores corresponding to `xvalues`.}

\item{dyi}{A numeric vector of standard errors for the `yvalues`.}

\item{s}{The fidelity (smoothing) constant for the spline.}

\item{xlow}{The index of the lowest score to be included in the spline fit.}

\item{xhigh}{The index of the highest score to be included in the spline fit.}

\item{ky}{The maximum possible score on the target scale (Y).}

\item{vectX}{The x-coordinates where the smoothed function should be evaluated.}
}
\value{
A list containing the smoothed y-values (`vectY`) and the spline
  coefficient matrix (`cmat`).
}
\description{
Implements the post-smoothing method described in Kolen & Brennan (2004),
which involves fitting a cubic spline to a portion of the data and using
linear interpolation for the tails.
}
