% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CubicSpline.R
\name{sspline}
\alias{sspline}
\title{Calculate Coefficients for a Smoothing Cubic Spline}
\usage{
sspline(x, y, dyi, s)
}
\arguments{
\item{x}{A numeric vector of the x-coordinates of the data points.}

\item{y}{A numeric vector of the y-coordinates of the data points.}

\item{dyi}{A numeric vector of the standard deviations of the y-coordinates.}

\item{s}{A fidelity constant that controls the closeness of the spline to the data.}
}
\value{
A matrix containing the spline coefficients (a, b, c, d).
}
\description{
This is the core function for calculating the coefficients of a smoothing
cubic spline. It implements the algorithm described by Reinsch (1967) and
referenced in Kolen & Brennan (2004).
}
\details{
The function iteratively finds an optimal smoothing parameter `p` and solves
a system of linear equations to determine the spline coefficients.
}
